Added Scenes, more controls, crude sat display
A lot.
Mainly, I decided the Quality can be an Enum class with Low and High, I added Scenes which are basically sub-views for the 3DGlobeView to determine what and how to render depending on the current state. Default is a simple static Earth globe, the others are one for tracking a current_satellite Satellite (defaults now to the ISS) with barely functioning position and camera following. I also added a scene state for the zoomed-out constellation overview, called Satellites Explorer Scene.